// data/github/current/ui/packages/item-picker2/projects/stories/__generated__/ItemPickerProjectsStoryQuery.graphql.ts
export type ItemPickerProjectsStoryQuery$variables = {
  number: number;
  owner: string;
  repo: string;
};
export type ItemPickerProjectsStoryQuery$data = {
  readonly repository: {
    readonly issue: {
      readonly projectCards: {
        readonly " $fragmentSpreads": FragmentRefs<"ItemPickerProjects_SelectedClassicProjectCardsFragment">;
      };
      readonly projectsV2: {
        readonly " $fragmentSpreads": FragmentRefs<"ItemPickerProjects_SelectedProjectsV2Fragment">;
      };
    } | null | undefined;
  } | null | undefined;
};
export type ItemPickerProjectsStoryQuery = {
  response: ItemPickerProjectsStoryQuery$data;
  variables: ItemPickerProjectsStoryQuery$variables;
};